DDP-7047 | eliminate code duplication by introducing template method in parser

+++ src/test/java/org/broadinstitute/dsm/model/elastic/export/MappingGeneratorTest.java
@@ -42,7 +42,7 @@ public class MappingGeneratorTest {
         Map<String, Object> objectMap = generator.generate(nameValue);
         Assert.assertEquals(objectMap.keySet().stream().findFirst().get(), BaseGenerator.PROPERTIES);
         String type = extractDeepestLeveleValue(objectMap);
-        Assert.assertEquals("integer", type);
+        Assert.assertEquals("long", type);
     }
 
     @Test
